+++ I2CQ73_HFT_2022231.Logic/Classes/TeamLogic.cs
@@ -23,17 +23,38 @@ namespace I2CQ73_HFT_2022231.Logic
 
 		public void Create(Team item)
 		{
-			this.teamRepo.Create(item);
+			if (item.TeamName.Length < 3)
+			{
+				throw new ArgumentException("The name is too short.");
+			}
+			else if (item.TeamName.Length > 240)
+			{
+				throw new ArgumentException("The name is too long.");
+			}
+			else
+			{
+				this.teamRepo.Create(item);
+			}
 		}
 
 		public void Delete(int id)
 		{
+			var car = this.teamRepo.Read(id);
+			if (car == null)
+			{
+				throw new ArgumentException("Team do not exists!");
+			}
 			this.teamRepo.Delete(id);
 		}
 
 		public Team Read(int id)
 		{
-			return this.teamRepo.Read(id);
+			var car = this.teamRepo.Read(id);
+			if (car == null)
+			{
+				throw new ArgumentException("Team do not exists!");
+			}
+			return car;
 		}
 
 		public IQueryable<Team> ReadAll()
@@ -43,7 +64,14 @@ namespace I2CQ73_HFT_2022231.Logic
 
 		public void Update(Team item)
 		{
-			this.teamRepo.Update(item);
+			if (item == null)
+			{
+				throw new NullReferenceException("Team do not exists!");
+			}
+			else
+			{
+				this.teamRepo.Update(item);
+			}
 		}
 	}
 }
